refactor: reorganize models
