+++ pkg/model/node_test.go
@@ -1,4 +1,4 @@
-package main
+package model
 
 import (
 	"testing"
