+++ pkg/model/utils.go
@@ -1,4 +1,4 @@
-package main
+package model
 
 import (
 	"strings"
