+++ pkg/model/wrap.go
@@ -1,4 +1,4 @@
-package main
+package model
 
 import (
 	"unicode/utf8"
